refactor(settings-button): Merge menu component into settings component
